+++ tests/controllers/person/list/index.js
@@ -1,0 +1,17 @@
+'use strict';
+var personList = {};
+
+module.exports = personList;
+
+personList.get = function*() {
+    // GET https://api.foo.com/v2/person/list
+    // Access models through this.models
+    this.data.people = [
+      {
+        name: "Jane",
+      },
+      {
+        name: "Bob"
+      }
+    ]
+};