+++ tests/createjwt.js
@@ -1,6 +1,7 @@
 var jwt = require('jwt-simple');
 
 var payload = {
+    sub: 'test',
     iat: new Date() / 1000,
     test: 'test'
 }