+++ tests/server.js
@@ -9,15 +9,11 @@ co(function* () {
         debug: true,
         tls: false,
         port: 8080,
-        //apiBase: 'https://api.foo.com/v2/',
-        // pass in the path to the raml file
-        raml: './raml/api_specification.raml',
-        controllerPath: '/controllers',
-        // pass in any models to make them easy to use
+        apiBase: 'https://api.foo.com/v2/',
+        raml: './tests/raml/api_specification.raml',
+        controllerPath: '/tests/controllers',
         models: {},
-        // koa-api-framework will automatically decode Authorization: Bearer [jwt] for you
         jwtSecret: 'qwertyuiopasdfghjklzxcvbnm123456',
-        // set which traits will decode the JWT
         authTraits: ['authenticated', 'administrator']
     });